[build] fixes for MSBuild project timing
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/plot/Build%20times/

The first issue with the new plots is I am storing `TimeSpan` data in
the CSV file. The Jenkins plugin for plots, doesn't seem to be able to
render this. I have converted the data to milliseconds instead.

The second issue is all of the `Install` targets for
`HelloWorld.csproj` are getting `[INSTALL_FAILED_NO_MATCHING_ABIS]`.
The fix for this is to pass `/p:AndroidSupportedAbis=x86` for each
build.