[build] fixes for MSBuild project timing (#1640)
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/plot/Build%20times/

Commit b0efc6ad added support for plotting build and install times.

The first issue with the new plots is we are storing `TimeSpan` data
in the CSV file.  The Jenkins plugin for plots, doesn't seem to be
able to render this.  We have converted the data to milliseconds
instead.

The second issue is all of the `Install` targets for
`HelloWorld.csproj` are getting `[INSTALL_FAILED_NO_MATCHING_ABIS]`.
The fix for this is to pass `/p:AndroidSupportedAbis=x86` for each
build.