[CodeBehind] Make sure the generated binding is not linked out
When the application with generated code-behind bindings is built in the Release
configuration with binding enabled (the default), the binding classes might be
linked out as they are instantiated using `Activator.CreateInstance` and not
directly. Prevent this from happening by decorating the binding constructor with
the `[Preserve]` custom attribute which tells the linker to keep the method in
the application.