[CodeBehind] Generate codebehind source in design time build
IDEs don't build the project until the user tells them to (either to debug or
just build) and thus initially the codebehind sources are not generated. This
causes the lack of intellisense as well as error reports from the IDE about
missing types etc. However, IDE uses a so-called "design-time build" to update
project resources and generate source files related to this task. Layout
codebehind generation should be done at the same time and this patch is an
attempt to make it so by adding the `_IncludeLayoutBindingSources` target
dependency to the `_ManagedUpdateAndroidResgen` target used by
design-time-build.