[Xamarin.Android.Build.Tasks] Intelligently update the `$(Intermediate)res` directory. (#1660)
For a long time it seems we have not been correctly updating the
`$(IntermediateOutputPath)res` directory from the
`_GenerateAndroidResourcesDir` target.  This has recently come to
light with our switch over to `msbuild` as our main way to compile
rather than `xbuild`.

For Target `Inputs` and `Outputs`, `msbuild` only checks the
timestamps of files to decide if they have changed.  However the
`<Copy/>` task checks *both* the timestamps *and the contents*.
It seems to do this even if the timestamps are the same or the
destination file is newer.  If the contents differ the file is copied.

This causes us a problem, since we fixup the destination files by
changing the cases of Resources.  Also we replace references to
managed class names with the md5 hashes Java class.  As a result if
we run the `_GenerateAndroidResourcesDir` target for any reason,
almost ALL the items will get updated. This will then cascade and
result in a new `.apk` being built.

This might explain some of the odd behaviour we get reported by our
uses regarding resources not being found, or long deployments.

To fix this, use our `<CopyIfChanged/>` task rather than `<Copy/>`.
This task only tries to copy the file IFF the source file is newer.
After that if the file did not change it will be skipped, but if it
did change it will be copied.  We also update the `<Touch/>` call to
make sure we only touch the resource flag file IF and ONLY IF:

 1. The file does not exist
 2. One of the files has been modified.

Some of the tests which wanted to force a resource build had to be
updated since they were just modifying the timestamp of the first
resource item in the project.  This generally is a drawable and never
changes so we never hit the problem where the source and destination
files are different.  As a result the expected targets did not run
because the system correctly detected no changes.