[Xamarin.Android.Build.Tasks] Code-Behind & design time build (#1655)
IDEs don't build the project until the user tells them to (either to
debug or just build) and thus initially the code-behind sources are
not generated.  This causes the lack of IntelliSense as well as error
reports from the IDE about missing types, etc.

However, Visual Studio uses "design-time build" to update project
resources and generate source files related to this task.  Layout
code-behind generation should be done at the same time and this patch
does so by adding the `_IncludeLayoutBindingSources` target
dependency to the `_ManagedUpdateAndroidResgen` target used by
design-time-builds.