[xabuild.exe] fixes for paths/settings on macOS
When looking at my Mono 5.12 installation, I noticed a few files we
should be using with `xabuild.exe`:
- `/Library/Frameworks/Mono.framework/Versions/5.12.0/lib/mono/msbuild/15.0/bin/NuGet.targets`
- `/Library/Frameworks/Mono.framework/Versions/5.12.0/lib/mono/msbuild/15.0/bin/Sdks` directory

`NuGet.targets` is needed for `XABuildPaths.NuGetRestoreTargets`, and
the `Sdks` directory can be used for `MSBuildSdksPath`.

Both of these values can still fall back to .NET Core, if they are not
found.

After these changes, it more closely matches the Windows code. We
should be able to build a wider range of .NET standard projects with
`xabuild.exe` now.