Fix bundled app crash on startup
Fixes: https://github.com/xamarin/xamarin-android/issues/1651

Recent update to mono brought in changes in the code generated by `mkbundle`
when dealing with AOT. Unfortunately, the changes require the bundle to be able
to find the `mono_jit_set_aot_mode` runtime function which isn't exported in the
Xamarin.Android up and thus the loading of the bundle crashes on startup with
the following message:

    dlopen failed: cannot locate symbol "mono_jit_set_aot_mode" referenced by libmonodroid_bundle_app.so

To fix this we need to augment the bundle's `mono_mkbundle_init` function in
generated code to take a third parameter - pointer to the abovementioned
function. With the current code this requires a set of rather ugly changes, as
seen in this commit, but it makes the application not crash on start and is the
fastest way to fix the issue right now.

This is a short-term fix, a long-term one will be presented at a later time.