[Xamarin.Android.Build.Tasks] IntermediateOutputPath includes TFN (#1628)
Xamarin.Android heavily leverages generated files within
`$(IntermediateOutputPath)` to achieve incremental builds and avoid
repeated invocations of Android SDK tools.  Many of the artifacts
within `$(IntermediateOutputPath)` are tied to
`$(TargetFrameworkVersion)` -- the API level that the project targets
-- which need to be regenerated whenever the value changes.

It is not uncommon for developers to frequently change
`$(TargetFrameworkVersion)`, e.g. to try out different API levels and
see which best fits their needs, and the subsequent
`$(IntermediateOutputPath)` invalidation can "needlessly" increase
build times if the `$(TargetFrameworkVersion)` is later changed to a
previously used (and generated) value.

In order to preserve the build artifacts which are generated,
override the `$(IntermediateOutputPath)` value so that it includes
`$(TargetFrameworkVersion)`, resulting in a value such as:

        obj\Debug\MonoAndroid81

`$(IntermediateOutputPath)` is only overridden in this manner when
`$(AndroidUseLatestPlatformSdk)`=False (which will become the default
in Visual Studio 15.8), as `$(IntermediateOutputPath)` needs to be
available before the `<ResolveSdks>` task can execute (which will
*update* `$(TargetFrameworkVersion)` when
`$(AndroidUseLatestPlatformSdk)`=True within the
`_SetLatestTargetFrameworkVersion` target).