[msbuild] Now JavaDoc task takes reference jars, and relevant tasks don't break.
Those changes were required to make it pass for building new androidx
libraries (but not limited to them).
Without references as in `javadoc -cp` they fail to resolve references.

It's okay, but JavaToolTask picks up all those messages as errors, and
the error code is respected and thus the entire binding library build fails.

That's not what we want. We should skip those failures and go on.

The subsequent crash at javadoc-to-mdoc failure needs investigation, but
it shouldn't block binding builds either.