[Xamarin.Android.Build.Tasks] ILRepack the NuGet.* Tasks Dependencies.
`ResolveAssemblies` is throwing the following error on
a newer mono (2018-04).

	error : Could not load file or assembly 'NuGet.ProjectModel, Version=4.6.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35' or one of its dependencies.

This is because we took a dependency on NuGet.ProjectModel 4.6 and
the newer mono has 4.7. This results in the error above.

Now we cannot always be sure which version of the NuGet assemblies
will be installed since users are able to upgrade mono independently.
So this PR will use the `ILRepack` NuGet to package up the dependencies
we need so that `Xamarin.Android.Build.Tasks` contains all the
correct assemblies it needs.