[Xamarin.Android.Build.Tasks] Fix bundled app crash on startup (#1683)
Fixes: https://github.com/xamarin/xamarin-android/issues/1651

Recent update to mono (166e7d52) brought in changes in the code
generated by `mkbundle.exe`, which is used when
`$(BundleAssemblies)`=True.  Unfortunately, the changes require the
bundle to be able to find the `mono_jit_set_aot_mode()` mono
embedding function, which doesn't work because
`libmonodroid_bundle_app.so` doesn't link against
`libmonosgen-2.0.so` -- because it can't, because on Android e.g.
`/data/data/@PACKAGE_NAME@/lib` isn't in `$LD_LIBRARY_PATH`, meaning
libraries "next to" `libmonodroid_bundle_app.so` won't be found by
**dlopen**(3), which is a constant source of ironically great fun --
which means that when we *do* try to call
`dlopen("libmonodroid_bundle_app")`, *it fails*:

        dlopen failed: cannot locate symbol "mono_jit_set_aot_mode" referenced by libmonodroid_bundle_app.so

To fix this we need to augment the bundle's `mono_mkbundle_init()`
function (which comes in via generated code) to take a third
parameter: a function pointer to `mono_jit_set_aot_mode`.  With the
current code this requires a set of rather ugly changes, but it makes
the application not crash on start and is the fastest way to fix the
issue right now.

This is a short-term fix.  A longer-term fix is being worked on:

        https://github.com/mono/mono/pull/8742

Additionally, update `RunTests.targets` so that we build
`src/Mono.Android/Test` with `$(BundleAssemblies)`=True and execute
the resulting `Mono.Android_Tests-Signed.apk` on the emulator to
ensure that `$(BundleAssemblies)`=True support works into the future.