Enabled DefaultOutputPaths targets for msbuild only
xbuild is not being compatible with the property evaluations
we're doing in the .targets => so we're only offering the output
paths improvements to the users that are using MsBuild.

It also prevents that CI builds don't fail because of this change.

And also adding the ability to disable it by setting the
EnableDefaultOutputPaths property to false.