Enabled DefaultOutputPaths targets for msbuild only (#1697)
xbuild is not being compatible with the property evaluations
we're doing in the .targets => so we're only offering the output
paths improvements to the users that are using MsBuild.

The problem is that `xbuild`, when using the new target is appending an extra `/` to the `$(IntermediateOutputPath)`. So we get `obj//Debug/foo` instead of `obj/Debug/foo`. This then causes problems later in the build.

Specifically in `ResolveLibraryProjectImports` we use `Directory.GetFiles`. When we pass in `obj//Debug/foo` it throws the following exception.

    ResolveLibraryProjectImports: System.ArgumentException: Second path fragment must not be a drive or UNC name

so because of the `//` it thinks the path is a UNC name. Now we could hack `ResolveLibraryProjectImports` to clean up the paths.. but there is no telling how many other places we would need to do that.


This commit also adds the ability to disable the new path format by setting the
EnableDefaultOutputPaths property to false.