[macOS] Drop requirement for Xcode when building Mono runtime
Mono SDK build currently checks for presence of Xcode when building on macOS.
However, Xcode isn't required to build neither Mono (for desktop or Android) nor
Xamarin.Android itself. This commit drops the dependency when building the
Android toolchains. The only build requirement, as far as tools are concerned,
on macOS are the command line tools from Apple.