[Xamarin.Android.Build.Tasks] Run ResolveNativeLibrariesInManagedReferences Test in Release
The `ResolveNativeLibrariesInManagedReferences` test fails when run
in Debug mode when we are using the `_BuildApkForFastDev`. This is
because the updated `libfoo.so` is not being updated in the apk.

Now this is probably a bug in the system. This is because the native
libraries are embedded into the apk even in fast dev. However to we
do not update the apk in fast dev mode if an assembly changes. This
is because we don't want to rebuild the apk every time a bit of
code is changed.

So this behaviour needs to be redone. This is will be handled
in the new instant run support in PR #609. This new system will
fast deploy native libraries to the device.

So for now so we can get a xamarin-adroid bump in monodroid we
should restrict this test to Release only.