[build, macOS] Drop requirement for Xcode when building Mono runtime (#1703)
Mono SDK build currently checks for presence of Xcode when building
on macOS.  However, Xcode isn't required to build both Mono (for
desktop or Android) or Xamarin.Android itself.

Drops the Xcode dependency when building the Android toolchains.  The
only build requirement, as far as tools are concerned, on macOS are
the command line tools from Apple.