[build] Fix build config mappings and project type guids. (#1708)
(You should not use MSBuild solution format unless everything can be
handled from within the IDEs or you understand what you are editing.)

Those ProjectTypeGuids were wrong; whoever made changes to the guids
should also fix Xamarin.Android.sln otherwise IDEs will fix them next
time someone with sanity makes changes on the file from IDEs and they
result in messy extraneous changes (and whoever used IDEs should not
be blamed; it is whoever didn't make required changes to blame).

Next, Java.Interop.Export was not appropriately added to configuration
mappings and therefore it was not built as part of dependencies when
it was built on the IDEs. `msbuild.exe` may build that but that's only
because build configuration is not taken into consideration.

This would not have happened if we have used some solution generator
instead of allowing bogus manual editing.