[Xamarin.Android.Build.Tasks] Fix up library project java classes (#1718)
Fixes: https://github.com/xamarin/xamarin-android/issues/1711

Commit a50e00bb removed one of the calls to the
`<ConvertResourcesCases/>` task.  This was responsible for the
library project layouts not being updated with the md5 generated java
class names.

However rather than adding the call back in we can expand the call in
the `_GenerateJavaStubs` target to include the library projects.
This makes more sense because that is the target in which we generate
the stubs in the first place.