[Xamarin.Android.Build.Tasks] <JavaDoc/> takes reference jars (#1687)
The `<JavaDoc/>` task now takes reference jars, and related tasks
don't break.  These changes are required to allow building the new
androidx libraries (but not limited to them).

Without references as in `javadoc -cp` they fail to resolve type
references.

It's okay, but `<JavaToolTask/>` picks up all the `javadoc` messages
as errors, and the error code is respected and thus the entire
binding library build fails.

That's not what we want.  We should skip those failures and go on.