[Xamarin.Android.Build.Tasks] Fix Intellisense errors for switch statements
We got a report that intellisense is showing errors on switch
statements for android resources. This might be because ALL of
the resource ids are 0 when using the managed parser.

This commit mimics the algorithm used by aapt as decribed at [1].
As a result most of the ids will be unique, which should fix
the intellisense errors.

[1] https://stackoverflow.com/questions/6517151/how-does-the-mapping-between-android-resources-and-resources-id-work