[tests] Set timeouts for Adb tasks
Set `Timeout` parameters for `Adb` tasks calls (and tasks based on
Adb).

We are still experiencing issues where `adb` get stuck for unknown
reason. Like in this case where it locked for nearly 7 hours.

```
01:39:56   adb I 05-25 01:39:56 71573 6448255 adb_io.cpp:75] readx: fd=3 wanted=4 (TaskId:88)
08:18:29 Build timed out (after 600 minutes). Marking the build as aborted.
```

We set the timeout in most cases to 1 or 2 minutes, plus these cases
with longer timeouts:

 * RunInstrumentationTests 30 minutes
 * RunUITests 5 minutes