[tests] Set timeouts for Adb tasks (#1733)
A [xamarin-android/master Jenkins build timed out][0], when
attempting to connect to the emulator (related: 284cd230, 3294a50e),
but this time with no obvious reason why it would fail.  

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/1024/

        01:39:36   Task StartAndroidEmulator (TaskId:86)
        01:39:36   Tool /Users/builder/android-toolchain/sdk/tools/emulator execution started with arguments: -verbose -avd XamarinAndroidTestRunner -port 5570 (TaskId:86)
                   # No apparent error messages
        01:39:56   /Users/builder/android-toolchain/sdk/platform-tools/adb -s emulator-5570 wait-for-device  (TaskId:87)
                   # Succeeded
        01:39:56   /Users/builder/android-toolchain/sdk/platform-tools/adb -s emulator-5570 shell 'counter=0; while [ $counter -lt 60 ] && [ "`getprop sys.boot_completed`" != "1" ]; do echo Waiting for device to fully boot; sleep 1; let "counter++"; done'  (TaskId:88)
        08:18:29 Build timed out (after 600 minutes). Marking the build as aborted.                   

None of the (voluminous) `adb` debug output was particularly useful
in suggesting *why* the last `adb shell` command just...hung.

To avoid 10+h build times that only end in a timeout, specify the
`ToolTask.Timeout` property on the `<Adb/>` task invocations, for
those which we feel reasonably confident in providing a timeout,
with most getting a timeout of 1 or 2 minutes, while
`RunInstrumentationTests` gets 30 minutes and `RunUITests` gets a
5 minute timeout.

This will hopefully avoid keeping the build server busy when the
build server isn't doing any useful work.