[docs] added ADB0000 error code
Context: https://github.com/xamarin/monodroid/pull/780
Context: https://github.com/xamarin/xamarin-android/issues/1560
Closes: https://github.com/xamarin/xamarin-android/issues/1727

I implemented `ADB0000` in the proprietary part of Xamarin.Android,
but need to document the error code in the open source
`xamarin-android` repository.

Looking at how the `Install` target is setup in `xamarin-android`, it
did not appear there was a good way to emit this error code for the
following target:

    <Target Name="_Deploy">
      <Exec Command="&quot;$(AdbToolPath)\adb&quot; install -r &quot;$(ApkFileSigned)&quot;" />
    </Target>

Since proprietary Xamarin.Android overrides this target, it is
probably not worth the work to write a custom MSBuild task to add the
`ADB0000` error code for open source `xamarin-android`.