[aapt2] _CreateAapt2VersionCache target should not run when aapt2 is disabled
Fixes: https://github.com/xamarin/xamarin-android/issues/1748

I was testing build times on master with a project in the
Xamarin.Forms repo.

I discovered the following scenario:
1. Build [Xamarin.Forms.ControlGallery.Android.csproj](https://github.com/xamarin/Xamarin.Forms/blob/master/Xamarin.Forms.ControlGallery.Android/Xamarin.Forms.ControlGallery.Android.csproj)
2. Build it again, with no changes

The `_CreateAapt2VersionCache` target was always running, and it seems
to be a bit of a performance regression.

For example, in the log from building this project I saw the following
times from this project and other referenced projects:
- 144ms
- 51ms
- 92ms
- 127ms
- 102ms

Total of 516ms added to a build with no changes.

To fix this I:
- Added a test with `$(AndroidUseAapt2)` set to `False`, verifying
  `Aapt2Link` and `_CreateAapt2VersionCache` are skipped.
- Added a `Condition` so that `_CreateAapt2VersionCache` doesn't run
  if aapt2 is disabled