[Linux] Fix a few problems with host build of libmonodroid
We've been building libmonodroid for Linux host incorrectly. While the shared
library built, it would never load properly, failing to resolve the
`pthread_threadid_np` and `print_ifla_name` functions.

The former is due the fact that we attempted to implement the `gettid`
function (returns current thread ID) which does not exist in glibc on Linux and
our implementation called `pthread_threadid_np` which caused a compiler warning
about missing declaration of the function - an error which would show up only at
runtime (but we practically don't use the Linux host libmonodroid currently and
so nobody noticed the issue). This commit fixes the implementation by calling
the `SYS_gettid` system call directly on Linux (but not on Android which has its
own implementation of the call in bionic)

The latter was due to missing `#ifdef DEBUG` around the call to the function
which is present only if `DEBUG` is defined.