[Xamarin.Android.Build.Tasks] fix duplicate Aapt2 warnings
When reviewing `ResolveSdksTask`, it appeared that failures to parse
`aapt2`'s version number would emit multiple warnings. The
`GetAapt2Version` method should only return `true` or `false`, since
the caller will emit its own `XA0111` warning. It was also odd that
the warnings generated in `GetAapt2Version` were coded to `XA0033` and
`XA0034`, since these are Java version warnings.

I also included the result of the `aapt2 version` call as a debug log
message. Since we would otherwise lose this information from a warning
that was removed.

I expanded the documentation about the following warnings:
- XA0111 - unknown aapt2 version
- XA0112 - aapt2 is missing

I did some testing with `$(AndroidUseAapt2)` enabled, checking the
behavior of these warnings.

When I rename `aapt2.exe` on my system, the following command:

    bin\Debug\bin\xabuild samples\HelloWorld\HelloWorld.csproj /p:AndroidUseAapt2=True

Results with a warning, but succeeds:

    warning XA0112: `aapt2` is not installed. Disabling `aapt2` support. Please check it is installed correctly.

This seems to be the desired scenario since the app continues to build
with plain `aapt`.