[monodroid] Fix the Linux host build (#1754)
We've been building `libmono-android.debug.so` for the Linux host
incorrectly.  While the shared library built, it would never load
properly, failing to resolve the `pthread_threadid_np()` and
`print_ifla_name()` functions.

`pthread_threadid_np()` couldn't be resolved due the fact that we
attempted to implement the [`gettid()`][0] function (returns current
thread ID) which does not exist in glibc on Linux and our
implementation called `pthread_threadid_np()` which caused a compiler
warning about missing declaration of the function:

        jni/monodroid-glue.c(1131,2): warning G500BB5CA: implicit declaration of function ‘pthread_threadid_np’; did you mean ‘pthread_tryjoin_np’? [-Wimplicit-function-declaration]
            pthread_threadid_np (NULL, &tid);
            ^~~~~~~~~~~~~~~~~~~

which turns into an error which would show up only at runtime (but we
practically don't use the Linux host `libmono-android.debug.so`
currently, so nobody noticed the issue).  This commit fixes the
implementation by calling the `SYS_gettid()` system call directly on
Linux, but not on Android which has its own implementation of the
call in Bionic.

`print_ifla_name()` was missing due to a missing `#ifdef DEBUG`
around the call to the function which is present only if `DEBUG`
is defined.

[0]: https://linux.die.net/man/2/gettid