[Xamarin.Android.ApiTools.ParameterNameExtractor] Enum Annotations (#325)
Add support for enum member annotations, newly introduced in API-P
stub sources.

The fixes also involves shift-reduce conflict changes related to
`package x.y.z` and `import x.y.z` where only dotted-identifiers are
expected.  They were found while I was fixing ambiguity on
Empty-or-enum-fields-and-then-class-members-in-enums part of the
construct (it was causing ambiguity and therefore adding annotations
on enum_identifier was not sufficient).